refactor(eol): split out battle code to `battle.cpp`
