+++ src/eol/battle.cpp
@@ -1,0 +1,161 @@
+#include "eol/eol.h"
+#include "eol/status_messages.h"
+#include "abc8.h"
+#include "level.h"
+#include "log.h"
+#include "pic8.h"
+#include <format>
+
+std::string eol::format_level(std::string_view level) {
+    std::string with_ext = std::format("{}.lev", level);
+    auto idx = get_internal_index(with_ext.c_str());
+
+    if (idx.has_value()) {
+        return std::format("internal {:02}", *idx);
+    }
+    return with_ext;
+}
+
+void eol::process(const battle_started& bs) {
+    current_battle = bs.bat;
+
+    if (current_battle->in_countdown) {
+        StatusMessages->add("battle countdown started");
+    } else {
+        StatusMessages->add("battle running");
+    }
+}
+
+void eol::process(const battle_countdown_ended&) {
+    if (!current_battle) {
+        LOG_ERROR("Received battle_countdown_ended message, but no battle is active");
+        return;
+    }
+
+    current_battle->in_countdown = false;
+    current_battle->local_start_ms = get_milliseconds();
+    StatusMessages->add("battle running");
+}
+
+void eol::process(const battle_ended& be) {
+    current_battle.reset();
+    StatusMessages->add(be.aborted ? "battle aborted" : "battle over");
+}
+
+void eol::process(const battle_time_sync& bts) {
+    if (!current_battle) {
+        LOG_ERROR("Received battle_time_sync message, but no battle is active");
+        return;
+    }
+
+    current_battle->local_start_ms = bts.local_start_ms;
+}
+
+static std::string format_hms(int seconds) {
+    seconds = std::max(seconds, 0);
+    int hours = seconds / 3600;
+    int mins = (seconds / 60) % 60;
+    int secs = seconds % 60;
+
+    if (hours > 0) {
+        return std::format("{}:{:02}:{:02}", hours, mins, secs);
+    }
+
+    return std::format("{}:{:02}", mins, secs);
+}
+
+static std::string_view battle_type_prefix(BattleType t) {
+    using enum BattleType;
+
+    switch (t) {
+    case Normal:
+        return "";
+    case OneLife:
+        return "one-life ";
+    case FirstFinish:
+        return "first finish ";
+    case Slowness:
+        return "slowness ";
+    case Survivor:
+        return "survivor ";
+    case LastCounts:
+        return "last counts ";
+    case FinishCount:
+        return "finish-count ";
+    case HourTT:
+        return "1 hour TT ";
+    case FlagTag:
+        return "flag tag ";
+    case AppleCollect:
+        return "apple ";
+    case Speed:
+        return "speed ";
+    }
+    return "";
+}
+
+static std::string format_type_with_cripples(BattleType t, BattleAttributes::Kind attrs) {
+    using namespace BattleAttributes;
+
+    constexpr std::pair<Kind, std::string_view> cripples[] = {
+        {NoVolt, "no-volt "},
+        {NoTurn, "no-turn "},
+        {OneTurn, "one-turn "},
+        {NoBrake, "no-brake "},
+        {NoThrottle, "no-throttle "},
+        {AlwaysThrottle, "always-throttle "},
+        {OneWheel, "one-wheel "},
+        {Drunk, "drunk "},
+        {Multi, "multi "},
+    };
+
+    std::string out{battle_type_prefix(t)};
+
+    for (auto [flag, text] : cripples) {
+        if (attrs & flag) {
+            out += text;
+        }
+    }
+
+    out += "battle";
+
+    out.front() = static_cast<char>(std::toupper(static_cast<unsigned char>(out.front())));
+
+    return out;
+}
+
+void eol::render_battle_status(pic8& dest, abc8& font) const {
+    if (!EolSettings->show_battle_status() || !current_battle) {
+        return;
+    }
+
+    const std::string type_text =
+        format_type_with_cripples(current_battle->type, current_battle->attributes);
+    const std::string level_text = format_level(current_battle->level_filename);
+    const std::string duration_text = format_hms(current_battle->duration * 60);
+    const std::string_view designer = lookup_nick(current_battle->designer_id);
+
+    const long long target_ms =
+        current_battle->in_countdown
+            ? current_battle->local_start_ms
+            : current_battle->local_start_ms + current_battle->duration * 60000LL;
+    const long long left_ms = target_ms - get_milliseconds();
+    const std::string time_text = format_hms(static_cast<int>((left_ms + 999) / 1000));
+
+    std::string out;
+    if (current_battle->in_countdown && current_battle->type == BattleType::FlagTag) {
+        out = std::format("{} in {} by {} - flag will be given in {} ({})", type_text, level_text,
+                          designer, time_text, duration_text);
+    } else if (current_battle->in_countdown) {
+        out = std::format("{} in {} by {} starts in {} ({})", type_text, level_text, designer,
+                          time_text, duration_text);
+    } else if (current_battle->type == BattleType::HourTT) {
+        out = std::format("{} by {} ends in {}", type_text, designer, time_text);
+    } else {
+        out = std::format("{} in {} by {} ends in {} / {}", type_text, level_text, designer,
+                          time_text, duration_text);
+    }
+
+    const int y = 15 + font.line_height() * (1 + EolSettings->chat_lines());
+    font.write_centered(&dest, dest.get_width() / 2, y, out.c_str());
+}
